+++ libControl/SliderVideo.cpp
@@ -93,7 +93,7 @@ void CSliderVideo::DrawShapeElement(wxDC * dc, const wxRect &rc)
 
 	if (secondTotalTime > 0)
 	{
-		pourcentage = (float)secondTimePast / (float)secondTotalTime;
+		pourcentage = secondTimePast / secondTotalTime;
 		rcPast.width = (rc.width * pourcentage);
 		rcPast.y = rc.y;
 		rcPast.height = themeSlider.GetRectangleHeight();
@@ -390,7 +390,7 @@ void CSliderVideo::CalculPositionButton()
 {
 	if (secondTotalTime > 0)
 	{
-		float pourcentage = (float)secondTimePast / (float)secondTotalTime;
+		float pourcentage = secondTimePast / secondTotalTime;
 		CalculPositionButton(positionSlider.x + int((float)(positionSlider.width) * pourcentage));
 	}
 	else
